+++ google_images_search/google_api.py
@@ -77,30 +77,39 @@ class GoogleCustomSearch(object):
         search_params = self._search_params(params)
         res = self._query_google_api(search_params, cache_discovery)
 
-        for image in res.get('items', []):
+        results = res.get('items', [])
+        if not results:
+            self._fetch_resize_save.zero_return = True
+
+        for image in results:
+
+            if len(self._fetch_resize_save._search_result) >= \
+                    self._fetch_resize_save._number_of_images:
+                break
+
             if self._fetch_resize_save.validate_images:
                 try:
                     response = requests.head(image['link'], timeout=5)
                     content_length = response.headers.get('Content-Length')
+                    content_type = response.headers.get('Content-Type', '')
 
                     # check if the url is valid
                     if response.status_code == 200 and \
-                            'image' in response.headers['Content-Type'] and \
-                            content_length:
+                            'image' in content_type and content_length:
+
                         # calculate download chunk size based on image size
                         self._fetch_resize_save.set_chunk_size(
                             image['link'], content_length
                         )
-
-                        # if everything is ok, yield it out
-                        yield image['link']
-
                 except requests.exceptions.ConnectTimeout:
                     pass
                 except requests.exceptions.ReadTimeout:
                     pass
                 except requests.exceptions.SSLError:
                     pass
+
+            yield image['link']
+
 
 
 class GoogleBackendException(Exception):